Move InspectorContextReduxAdapter into its own file to fix import issue
